chore: renamed project
